Supress #10: AbortError: This operation was aborted
After several days of debugging, we have come to the conclusion that we're unable to patch this bug. To avoid it from constantly bugging devs whenever it triggers, This commit adds ignored errors to the Pager, which identifies an error via a 64bit hash generated by the stack trace. If an error hash matches an ignored error, it is dropped. It will still be logged to the console, however.

Future errors and crashes will also show it's hash on the paged message.